Killer-Specific strategy
Implemented a way for strategies to only apply to specific constraints.
Added a Killer Sudoku strategy that lists all possibilities for cages
and removes all options that do not occur in one of those. This yields a
huge performance improvement and should be sufficient to accept Killer
Sudoku into the benchmark suite soon.
Renamed the `impls` module in `strategy` to `general`. It now only
contains strategies which apply to all constraints.
Refactored index computation to be nicer.
Processed a TODO in SudokuInfo for better style.